Support Multiple Targets

+++ get_openssl.sh
@@ -3,19 +3,25 @@ set -e
 function get {
   if [ ! -d "$ourname" ]
   then
-    if [ ! -f "$tarname" ]
+    if [ ! \( -f $tarname -o -L $tarname \) ]
     then
-      if [ -f "../$tarname" ]
+      if [ -f ../$tarname -o -L ../$tarname ]
       then
         ln -s "../$tarname" "$tarname"
+      elif [ -f ../../$tarname -o -L ../../$tarname ]
+      then
+        ln -s "../../$tarname" "$tarname"
       else
         curl -OJL "$url"
       fi
     fi
 
-    tar -xzf "$tarname"
-    mv "$basename" "$ourname"
-    rm -f "$tarname"
+    if ! ${JUST_CACHE_SOURCES:-false}
+    then
+      tar -xzf "$tarname"
+      mv "$basename" "$ourname"
+      rm -f "$tarname"
+    fi
   fi
 }
