+++ get_xerces.sh
@@ -3,19 +3,25 @@ set -e
 function get {
   if [ ! -d "$ourname" ]
   then
-    if [ ! -f "$tarname" ]
+    if [ ! \( -f $tarname -o -L $tarname \) ]
     then
-      if [ -f "../$tarname" ]
+      if [ -f ../$tarname -o -L ../$tarname ]
       then
         ln -s "../$tarname" "$tarname"
+      elif [ -f ../../$tarname -o -L ../../$tarname ]
+      then
+        ln -s "../../$tarname" "$tarname"
       else
         curl -OJL "$url"
       fi
     fi
 
-    tar -xzf "$tarname"
-    mv "$basename" "$ourname"
-    rm -f "$tarname"
+    if ! ${JUST_CACHE_SOURCES:-false}
+    then
+      tar -xzf "$tarname"
+      mv "$basename" "$ourname"
+      rm -f "$tarname"
+    fi
   fi
 }
 
@@ -32,6 +38,11 @@ tarname="$basename.tar.gz"
 url="http://apache.cs.utah.edu/xerces/c/3/sources/$tarname"
 ourname="xerces_source"
 get
+
+if ${JUST_CACHE_SOURCES:-false}
+then
+  exit 0
+fi
 
 # Patch Xerces To Use Our Libiconv
 if ! md5sum -c "${basename}.md5" --status
